Rename decoration to `ListItemDividerDecoration`
