+++ android/src/main/kotlin/net/mullvad/mullvadvpn/ui/ListItemDividerDecoration.kt
@@ -1,4 +1,4 @@
-package net.mullvad.mullvadvpn.relaylist
+package net.mullvad.mullvadvpn.ui
 
 import android.content.Context
 import android.graphics.Rect
@@ -8,8 +8,8 @@ import android.support.v7.widget.RecyclerView.State
 import android.view.View
 import net.mullvad.mullvadvpn.R
 
-class RelayItemDividerDecoration(private val context: Context) : ItemDecoration() {
-    private val dividerHeight = context.resources.getDimensionPixelSize(R.dimen.relay_list_divider)
+class ListItemDividerDecoration(private val context: Context) : ItemDecoration() {
+    private val dividerHeight = context.resources.getDimensionPixelSize(R.dimen.list_item_divider)
 
     override fun getItemOffsets(offsets: Rect, view: View, parent: RecyclerView, state: State) {
         val position = parent.getChildAdapterPosition(view)
